Fix code style & class encapsulation
Task class description and isDone fields are currently public.

This violates encapsulation principles and exposes internal state
directly, making the class harder to maintain and prone
to unintended modifications. Additionally, code style was not
up to mark.

With this commit, the key changes are:
*   Make the description and isDone fields private.
*   Add getter methods for retrieving the values of these fields.
*   Add setter methods for modifying the values of these fields.
*  Amended all code style errors with checkStyle plugin

This improves encapsulation, provides controlled access to the
task's state, and increases the class's robustness. This commit also
fixes glaring code style issues to align with CS2103 coding
style standards for greater readability.