Amend README.md and HelpCommand to account for Priority
The user guide(UG)/README.md lacked documentation on the newly-introduced Priority feature for Tasks. This is
also the case for the HelpCommand.

This needs to change to ensure that the UG and HelpCommand is comprehensive and that new users are able to be
quickly onboarded by simply reading the UG.

Let's
- Append details regarding the implementation of the Priority system
- Provide details on how users can use the Priority system including syntax and potential errors
- Append similar details to HelpCommand.java

By doing so, this gives users enough clarity when reading the UG or
using the HelpCommand via 'help'.